+++ src/main/java/com/danrus/rpt/impl/conditional/InFluidProperty.java
@@ -1,0 +1,25 @@
+package com.danrus.rpt.impl.conditional;
+
+import com.mojang.serialization.MapCodec;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.renderer.item.properties.conditional.ConditionalItemModelProperty;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemDisplayContext;
+import net.minecraft.world.item.ItemStack;
+import org.jetbrains.annotations.Nullable;
+
+public record InFluidProperty() implements ConditionalItemModelProperty {
+
+    public static final MapCodec<InFluidProperty> MAP_CODEC = MapCodec.unit(new InFluidProperty());
+
+    @Override
+    public MapCodec<? extends ConditionalItemModelProperty> type() {
+        return MAP_CODEC;
+    }
+
+    @Override
+    public boolean get(ItemStack stack, @Nullable ClientLevel level, @Nullable LivingEntity entity, int seed, ItemDisplayContext displayContext) {
+        if (entity == null) return false;
+        return !entity.fluidOnEyes.isEmpty();
+    }
+}