+++ src/main/java/com/danrus/rpt/mixin/regs/ConditionalItemModelPropertiesMixin.java
@@ -1,6 +1,8 @@
 package com.danrus.rpt.mixin.regs;
 
+import com.danrus.rpt.impl.conditional.EntityFlagProperty;
 import com.danrus.rpt.impl.conditional.HasFlagProperty;
+import com.danrus.rpt.impl.conditional.InFluidProperty;
 import com.danrus.rpt.impl.conditional.MatchCustomNameRegexProperty;
 import com.mojang.serialization.MapCodec;
 import net.minecraft.client.renderer.item.properties.conditional.ConditionalItemModelProperties;
@@ -28,5 +30,7 @@ public class ConditionalItemModelPropertiesMixin {
     private static void rpt$injectBootstrap(CallbackInfo ci) {
         ID_MAPPER.put(ResourceLocation.fromNamespaceAndPath("rpt", "has_flag"), HasFlagProperty.MAP_CODEC);
         ID_MAPPER.put(ResourceLocation.fromNamespaceAndPath("rpt", "match"), MatchCustomNameRegexProperty.MAP_CODEC);
+        ID_MAPPER.put(ResourceLocation.fromNamespaceAndPath("rpt", "in_fluid"), InFluidProperty.MAP_CODEC);
+        ID_MAPPER.put(ResourceLocation.fromNamespaceAndPath("rpt", "entity_flag"), EntityFlagProperty.MAP_CODEC);
     }
 }